+++ 1.21.5/Common/src/main/java/fuzs/puzzleslib/api/client/core/v1/ClientModConstructor.java
@@ -96,7 +96,7 @@ public interface ClientModConstructor {
 
     /**
      * @param context register a resolver responsible for mapping each {@link BlockState} of a block to an
-     *                {@link net.minecraft.client.renderer.block.model.UnbakedBlockStateModel}
+     *                {@link net.minecraft.client.renderer.block.model.BlockStateModel.UnbakedRoot}
      */
     default void onRegisterBlockStateResolver(BlockStateResolverContext context) {
         // NO-OP