+++ 1.21.5/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.5.1-1.21.5] - 2025-04-01
+### Changed
+- Overhaul internal development structure
+### Fixed
+- Fix client start-up crash on NeoForge caused by `Minecraft` being `null`
+
 ## [v21.5.0-1.21.5] - 2025-03-31
 - Port to Minecraft 1.21.5
 ### Added